Add temporary discounts on fruit and vegetables
Added discount tags to product cards that belong to the fruit and
vegetables category. Old price is crossed out and the price after the
discount is displayed. In cart page, the bill now takes into account the
discounted amount to be subtracted from the subtotal. Only made changes
concerning product card, logged in user cart and bill.