Fix cart bug
When user is logged in and cart is refreshed the local cart is displayed
then the remote cart is rendered instead. This was because upon loading
the app the wrong token variable was checked, and the tags attribute for
localcart items is an object rather than an array, whereas for the
remotecart items it's an array, creating a conflict when accessing the
same attribute. Now if userToken does not exist, the local cart is saved to state and if it does the remote cart
is. This magically fixed the create order function.